+++ frontend/src/pages/helpcenter/station/setup/federation.vue
@@ -4,14 +4,14 @@
  *     Copyright (C) RainbowDashLabs and Contributor
  */
 <script setup lang="ts">
-import OverviewHelp from '~/views/helpcenter/stationview/setup/OverviewHelp.vue'
+import FederationStepHelp from '~/views/helpcenter/stationview/setup/FederationStepHelp.vue'
 
 definePageMeta({
   layout: 'helpcenter',
-  name: 'help-setup-module-overview',
+  name: 'help-station-setup-federation',
 })
 </script>
 
 <template>
-  <OverviewHelp/>
+  <FederationStepHelp/>
 </template>