+++ frontend/src/pages/helpcenter/station/setup/groups.vue
@@ -4,14 +4,14 @@
  *     Copyright (C) RainbowDashLabs and Contributor
  */
 <script setup lang="ts">
-import OverviewHelp from '~/views/helpcenter/stationview/setup/OverviewHelp.vue'
+import GroupsStepHelp from '~/views/helpcenter/stationview/setup/GroupsStepHelp.vue'
 
 definePageMeta({
   layout: 'helpcenter',
-  name: 'help-setup-module-overview',
+  name: 'help-station-setup-groups',
 })
 </script>
 
 <template>
-  <OverviewHelp/>
+  <GroupsStepHelp/>
 </template>